backend/requirements.txt
azure-functions==1.18.0
requests==2.31.0
azure-storage-blob==12.19.0